requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.3
decorator==4.3.2
future==0.17.1
gast==0.2.2
glfw==1.7.0
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imageio==2.4.1
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
joblib==0.13.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
keyword2cmdline==1.0.2
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.0.1
matplotlib==3.0.2
mpi4py==3.0.0
mujoco-py==1.50.1.68
numpy==1.16.0
opencv-contrib-python==4.0.0.21
pandas==0.24.0
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
rl-algs==0.0.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.30.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
